+++ library/src/schemas/date/date.test.ts
@@ -11,12 +11,8 @@ describe('date', () => {
     expect(output).toEqual(input);
     expect(() => parse(schema, 2023)).toThrowError();
     expect(() => parse(schema, '2023-07-10')).toThrowError();
+    expect(() => parse(schema, new Date('Invalid Date'))).toThrowError();
     expect(() => parse(schema, {})).toThrowError();
-  });
-
-  test('should pass only valid dates', () => {
-    const error = "Date is invalid!";
-    expect(() => parse(date(error), new Date("Not a date"))).toThrowError(error);
   });
 
   test('should throw custom error', () => {